refactor(tocco-ui): delete FormField
- FormField was replaced by StatedValue
- move ErrorList from FormField to StatedValue
- all tests of FormField has a substitute in StatedValue
