--- packages/tocco-ui/src/FormField/index.js
@@ -1,3 +1,0 @@
-import FormField from './FormField'
-
-export default FormField
